requirements.txt
# AgenticX-GUIAgent 移动GUI智能体系统依赖
# 基于AgenticX框架的移动设备自动化操作系统
# 支持自然语言驱动的Android GUI操作

# ============================================================================
# AgenticX 核心框架
# ============================================================================
# 注意: AgenticX需要从本地安装 (pip install -e /path/to/AgenticX)
# agenticx>=0.1.0

# ============================================================================
# 异步支持和网络通信
# ============================================================================
aiohttp>=3.9.0
aiofiles>=23.2.0
asyncio-mqtt>=0.16.0
websockets>=12.0
httpx>=0.25.0

# ============================================================================
# 日志系统
# ============================================================================
loguru>=0.7.0

# ============================================================================
# 移动设备控制和自动化
# ============================================================================
# ADB和Android设备控制
appium-python-client>=4.0.0
selenium>=4.16.0
uiautomator2>=3.2.0
adbutils>=0.11.0
pure-python-adb>=0.3.0.dev0

# iOS设备支持（可选）
# wda>=1.4.0
# tidevice>=0.9.0

# ============================================================================
# 图像处理和计算机视觉
# ============================================================================
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
scikit-image>=0.22.0
imageio>=2.33.0

# OCR支持
pytesseract>=0.3.10
easyocr>=1.7.0

# ============================================================================
# 机器学习和深度学习
# ============================================================================
# PyTorch生态
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers和NLP
transformers>=4.36.0
sentence-transformers>=2.2.2
tokenizers>=0.15.0

# 多模态模型支持
openai-clip>=1.0.1
# timm>=0.9.0

# 向量数据库和相似度搜索
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # GPU版本（可选）

# ============================================================================
# LLM API客户端
# ============================================================================
openai>=1.6.0
anthropics>=0.8.0
# google-generativeai>=0.3.0

# ============================================================================
# 数据处理和分析
# ============================================================================
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ============================================================================
# 配置管理和序列化
# ============================================================================
pyyaml>=6.0.1
pydantic>=2.5.0
hydra-core>=1.3.2
python-dotenv>=1.0.0
toml>=0.10.2

# ============================================================================
# 日志和监控
# ============================================================================
loguru>=0.7.2
prometheus-client>=0.19.0
rich>=13.7.0
colorama>=0.4.6

# ============================================================================
# 数据库和存储
# ============================================================================
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
redis>=5.0.0

# ============================================================================
# Web框架和API
# ============================================================================
fastapi>=0.104.0
uvicorn>=0.24.0
starlette>=0.27.0

# ============================================================================
# 测试框架
# ============================================================================
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0

# ============================================================================
# 开发工具
# ============================================================================
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.8.0
pre-commit>=3.6.0

# ============================================================================
# 实用工具
# ============================================================================
requests>=2.31.0
click>=8.1.7
tqdm>=4.66.0
psutil>=5.9.6
typing-extensions>=4.8.0

# ============================================================================
# 时间和调度
# ============================================================================
schedule>=1.2.0
apscheduler>=3.10.0

# ============================================================================
# 安全和加密
# ============================================================================
cryptography>=41.0.0

# ============================================================================
# 可选依赖（根据需要安装）
# ============================================================================
# 语音识别和TTS
# speechrecognition>=3.10.0
# pyttsx3>=2.90
# azure-cognitiveservices-speech>=1.34.0

# 更多图像处理
# matplotlib>=3.8.0
# seaborn>=0.13.0

# 数据可视化
# plotly>=5.17.0
# dash>=2.14.0

# 机器人流程自动化
# pyautogui>=0.9.54
# pynput>=1.7.6

tensorboard
scipy
gym